Move Skeleton to the Display section
