+++ lib/components/display/skeleton.jsx
